+++ lib/client-api/src/story_store.ts
@@ -110,8 +110,6 @@ export default class StoryStore {
 
   _argTypesEnhancers: ArgTypesEnhancer[];
 
-  _revision: number;
-
   _selection: Selection;
 
   constructor(params: { channel: Channel }) {
@@ -123,7 +121,6 @@ export default class StoryStore {
     this._kinds = {};
     this._stories = {};
     this._argTypesEnhancers = [];
-    this._revision = 0;
     this._selection = {} as any;
     this._error = undefined;
     this._channel = params.channel;
@@ -521,14 +518,6 @@ export default class StoryStore {
 
   getRawStory(kind: string, name: string) {
     return this.getStoriesForKind(kind).find((s) => s.name === name);
-  }
-
-  getRevision() {
-    return this._revision;
-  }
-
-  incrementRevision() {
-    this._revision += 1;
   }
 
   cleanHooks(id: string) {