+++ lib/core/src/client/preview/loadCsf.test.ts
@@ -33,7 +33,6 @@ function makeMocks() {
   const configApi = ({ configure: (x: Function) => x() } as unknown) as ConfigApi;
   const storyStore = ({
     removeStoryKind: jest.fn(),
-    incrementRevision: jest.fn(),
   } as unknown) as StoryStore;
   const clientApi = ({
     storiesOf: jest.fn().mockImplementation(() => ({
@@ -323,7 +322,6 @@ describe('core.preview.loadCsf', () => {
     configure(makeRequireContext(secondInput), mod, 'react');
 
     expect(storyStore.removeStoryKind).not.toHaveBeenCalled();
-    expect(storyStore.incrementRevision).not.toHaveBeenCalled();
     expect(mockedStoriesOf).toHaveBeenCalledWith('b', true);
   });
 
@@ -357,7 +355,6 @@ describe('core.preview.loadCsf', () => {
     configure(makeRequireContext(secondInput), mod, 'react');
 
     expect(storyStore.removeStoryKind).toHaveBeenCalledWith('b');
-    expect(storyStore.incrementRevision).toHaveBeenCalled();
     expect(mockedStoriesOf).not.toHaveBeenCalled();
   });
 
@@ -400,7 +397,6 @@ describe('core.preview.loadCsf', () => {
 
     expect(storyStore.removeStoryKind).toHaveBeenCalledTimes(1);
     expect(storyStore.removeStoryKind).toHaveBeenCalledWith('a');
-    expect(storyStore.incrementRevision).toHaveBeenCalled();
     expect(mockedStoriesOf).toHaveBeenCalledWith('a', true);
   });